git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@15555 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/feature.h
@@ -52,7 +52,7 @@ extern MagickExport Image
     const double,ExceptionInfo *),
   *HoughLineImage(const Image *,const size_t,const size_t,const size_t,
     ExceptionInfo *),
-  *MeanShiftImage(const Image *,const size_t,const size_t,const size_t,
+  *MeanShiftImage(const Image *,const size_t,const size_t,const double,
     ExceptionInfo *);
 
 #if defined(__cplusplus) || defined(c_plusplus)
